+++ src/server/physics/physics-server.ts
@@ -19,6 +19,7 @@ import './physics-server-sender'
 import './storage/storage'
 import './disable-idle-pose'
 import './event/event'
+import './server-vars'
 
 export interface PhysicsServerConnectionSettings {
     httpPort: number